Split delta / message choice
These choices behave differently.
